Scanner v3/requirements.txt
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Cython==3.0.12
decorator==5.2.1
easyocr==1.7.2
filelock==3.18.0
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
fsspec==2025.3.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy_loader==0.4
lmdb==1.6.2
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.10.0
paddlepaddle==3.0.0
pillow==11.2.1
protobuf==6.30.2
pyclipper==1.3.0.post6
pydantic==2.11.4
pydantic_core==2.33.2
python-bidi==0.6.6
python-docx==1.1.2
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
setuptools==80.1.0
shapely==2.1.0
simsimd==6.2.1
sniffio==1.3.1
soupsieve==2.7
stringzilla==3.12.5
sympy==1.14.0
termcolor==3.1.0
tifffile==2025.3.30
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wheel==0.45.1
